+++ src/vs/workbench/contrib/chat/browser/chatContentParts/toolInvocationParts/chatToolInvocationSubPart.ts
@@ -1,0 +1,34 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { Emitter } from '../../../../../../base/common/event.js';
+import { Disposable } from '../../../../../../base/common/lifecycle.js';
+import { IChatToolInvocation, IChatToolInvocationSerialized } from '../../../common/chatService.js';
+import { IChatCodeBlockInfo } from '../../chat.js';
+
+export abstract class BaseChatToolInvocationSubPart extends Disposable {
+	protected static idPool = 0;
+	public abstract readonly domNode: HTMLElement;
+
+	protected _onNeedsRerender = this._register(new Emitter<void>());
+	public readonly onNeedsRerender = this._onNeedsRerender.event;
+
+	protected _onDidChangeHeight = this._register(new Emitter<void>());
+	public readonly onDidChangeHeight = this._onDidChangeHeight.event;
+
+	public abstract codeblocks: IChatCodeBlockInfo[];
+
+	public readonly codeblocksPartId = 'tool-' + (BaseChatToolInvocationSubPart.idPool++);
+
+	constructor(
+		toolInvocation: IChatToolInvocation | IChatToolInvocationSerialized,
+	) {
+		super();
+
+		if (toolInvocation.kind === 'toolInvocation' && !toolInvocation.isComplete) {
+			toolInvocation.isCompletePromise.then(() => this._onNeedsRerender.fire());
+		}
+	}
+}
